+++ kernels/cuda/sqrt.cu
@@ -1,10 +1,10 @@
 #include "../../include/utils.h"
 
 __global__ void
-square_root_newton_raphson (double *input, double *output, int n)
+square_root_newton_raphson (float *input, float *output, int n)
 {
   int tid = threadIdx.x + blockIdx.x * blockDim.x;
-  double approx = input[tid] * .5;
+  float approx = input[tid] * .5;
   if (input[tid] < 0)
     {
       return;
@@ -21,7 +21,7 @@ square_root_newton_raphson (double *input, double *output, int n)
 }
 
 __global__ void
-sqrt_libmath (double *input, double *output, int n)
+sqrt_libmath (float *input, float *output, int n)
 {
   int tid = threadIdx.x + blockIdx.x * blockDim.x;
     output[tid] = sqrt (input[tid]);
